feat(sugar): restructures `retry`
